Update main.c
All bugs in the game have been fixed.
Added functions for random generation of the game side and Y position of tetrominos.
Added a game over screen along with hearts representing 'lives' in the game. 
The game starts with three lives. 
The game speed and score change only when all lives are lost.